Move 'new connection' route to own folder
