+++ Source/SelfService/Web/integrations/connection/new/index.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import React from 'react';
-import { Page } from '../../components/layout/page';
+import { Page } from '../../../components/layout/page';
 
 export const NewConnection = () => {
